requirements.txt

python-dotenv
requests
pandas
chainlit
litellm